Separate timeslicing process into dependent module
