+++ lib/embulk/input/marketo/lead.rb
@@ -4,52 +4,14 @@ module Embulk
   module Input
     module Marketo
       class Lead < Base
+        include Timeslice
+
         PREVIEW_COUNT = 15
 
         Plugin.register_input("marketo/lead", self)
 
         def self.target
           :lead
-        end
-
-        def self.transaction(config, &control)
-          endpoint_url = config.param(:endpoint, :string)
-
-          if config.param(:last_updated_at, :string, default: nil)
-            Embulk.logger.warn "config: last_updated_at is deprecated. Use since_at/until_at"
-          end
-
-          task = {
-            endpoint_url: endpoint_url,
-            wsdl_url: config.param(:wsdl, :string, default: "#{endpoint_url}?WSDL"),
-            user_id: config.param(:user_id, :string),
-            encryption_key: config.param(:encryption_key, :string),
-            since_at: config.param(:since_at, :string),
-            until_at: config.param(:until_at, :string, default: Time.now.to_s),
-            columns: config.param(:columns, :array)
-          }
-
-          columns = []
-
-          task[:columns].each do |column|
-            name = column["name"]
-            type = column["type"].to_sym
-
-            columns << Column.new(nil, name, type, column["format"])
-          end
-
-          resume(task, columns, 1, &control)
-        end
-
-        def self.guess(config)
-          if config.param(:last_updated_at, :string, default: nil)
-            Embulk.logger.warn "config: last_updated_at is deprecated. Use since_at/until_at"
-          end
-
-          client = soap_client(config)
-          metadata = client.metadata
-
-          return {"columns" => generate_columns(metadata)}
         end
 
         def self.generate_columns(metadata)
